* move function from contacts.ts to browser.ts * include a usage explanation in browser.ts
